+++ halo2_proofs/src/plonk/verifier/batch.rs
@@ -1,16 +1,17 @@
-use std::{io, marker::PhantomData};
-
 use group::ff::Field;
 use pasta_curves::arithmetic::CurveAffine;
-use rand_core::{OsRng, RngCore};
-use rayon::iter::{IndexedParallelIterator, IntoParallelIterator, ParallelIterator};
+use rand_core::OsRng;
 
 use super::{verify_proof, VerificationStrategy};
 use crate::{
+//<<<<<<< HEAD
     //multicore, // mschoenebeck: commented out because it's unused(?)
+//=======
+    multicore::{IndexedParallelIterator, IntoParallelIterator, ParallelIterator},
+//>>>>>>> 6ae9f77e04d471c64b31b86486fb6ae974dc31a1
     plonk::{Error, VerifyingKey},
     poly::commitment::{Guard, Params, MSM},
-    transcript::{Blake2bRead, EncodedChallenge, TranscriptRead},
+    transcript::{Blake2bRead, EncodedChallenge},
 };
 
 /// A proof verification strategy that returns the proof's MSM.