+++ halo2_proofs/CHANGELOG.md
@@ -6,6 +6,27 @@ and this project adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- The following structs now derive the `Eq` trait:
+  - `halo2_proofs::dev`:
+    - `failure::FailureLocation`
+    - `failure::VerifyFailure`
+    - `metadata::Gate`
+    - `metadata::Constraint`
+    - `metadata::Region`
+  - `halo2_proofs::poly::Rotation`
+
+### Added
+- `halo2_proofs::arithmetic::FftGroup`
+- `halo2_proofs::circuit::layouter`:
+  - `RegionLayouter::instance_value` method added to provide access to
+    instance values within a region.
+
+### Changed
+- APIs with `F: pasta_curves::arithmetic::FieldExt` bounds have been changed to
+  use `ff` traits directly.
+- `halo2_proofs::arithmetic`:
+  - `best_fft, recursive_butterfly_arithmetic` now use the `FftGroup` trait
+    instead of the (now-removed) `pasta_curves::arithmetic::Group` trait.
 
 ## [0.2.0] - 2022-06-23
 ### Added